+++ lake/api/api.go
@@ -5,6 +5,7 @@ import (
 	"encoding/hex"
 	"fmt"
 
+	"github.com/brimdata/zed"
 	"github.com/brimdata/zed/api"
 	"github.com/brimdata/zed/lake"
 	"github.com/brimdata/zed/lake/index"
@@ -14,6 +15,7 @@ import (
 	"github.com/brimdata/zed/pkg/storage"
 	"github.com/brimdata/zed/zbuf"
 	"github.com/brimdata/zed/zio"
+	"github.com/brimdata/zed/zson"
 	"github.com/segmentio/ksuid"
 )
 
@@ -147,4 +149,26 @@ func LookupBranchByID(ctx context.Context, api Interface, id ksuid.KSUID) (*lake
 
 func idToHex(id ksuid.KSUID) string {
 	return hex.EncodeToString(id.Bytes())
+}
+
+type buffer struct {
+	unmarshaler *zson.UnmarshalZNGContext
+	results     []interface{}
+}
+
+var _ zio.Writer = (*buffer)(nil)
+
+func newBuffer(types ...interface{}) *buffer {
+	u := zson.NewZNGUnmarshaler()
+	u.Bind(types...)
+	return &buffer{unmarshaler: u}
+}
+
+func (b *buffer) Write(val *zed.Value) error {
+	var v interface{}
+	if err := b.unmarshaler.Unmarshal(*val, &v); err != nil {
+		return err
+	}
+	b.results = append(b.results, v)
+	return nil
 }
